Observer Pattern and Snipeshot
I now use a kind of Observer Pattern setup to manage all of the objects
that need to be updated and drawn. All objects that want to use it
request to have such via Game.addUpdateable() and can request to stop
being updated and drawn with Game.removeUpdateable(). I've also added a
basic form of the snipeshot implemented. Nothing for it to shoot,
though.